Remove number keys from dvorak layer
The standard dvorak layer had number keys in the top row of it, this
lead to problems because the top row is so far away from the homerow. To
get used to number keys on the function layer, the dvorak layer number
keys have been removed.

Additionally some cleanup has been done and a script for automatically
installing everything required on arch + updating the firmware has been
added.