Add Copy/Paste keys
Since copy/pasting with Dvorak and one hand on the mouse can be a bit of
a pain sometimes, an easy shortcut for Copy/Paste keys has been added
which can be executed without using the right hand.

This also moves away the RESET functionality a bit more to make sure
it's never, ever hit again accidentally.