Add Alacritty logo animations
This adds a simple animation to the Alacritty logo which switches
between a big and small flame in the logo with every key press.

Additionally this also adds an animation during startup which will move
the Alacritty logo from the bottom of the screen to the top.

To make updating the firmware a little easier should the reset button
ever be occluded, a dedicated binding for a reset key has also been
added to the keymap.