Add deadzone to maximized joystick values
This fixes an issue where it was difficult to get the maximum range of
motion out of joysticks since the sensors didn't manage to reliably
report the maximum value when fully extended in a certain direction.

To fix this a deadzon has been added to the outside, in addition to the
deadzone around the joysticks center. As a result values inside the
outside deadzone are now interpreted as 100%, similar to how value
sinside the central deadzon are considered 0%.

This means there's less range of motion available in the joysticks
themselves, but the reliability should be more valueable in this case.
Between the deadzones the speed is still just linearly interpolated.